refactor: merge pending transfers and multisignatures into new `PendingTransactionsTable`
