+++ src/domains/transaction/components/TransactionTable/PendingTransactionsTable/index.tsx
@@ -1,0 +1,1 @@
+export * from "./PendingTransactionsTable";
